move mock message value to `test/values`
