+++ src/home/cleverTapInbox.ts
@@ -103,62 +103,3 @@ interface Media {
   key?: string
   processing?: boolean
 }
-
-export const mockCleverTapInboxMessage = {
-  wzrkParams: { wzrk_id: '0_0' },
-  id: '1704393845',
-  wzrk_id: '0_0',
-  msg: {
-    tags: [],
-    type: 'message-icon',
-    content: [
-      {
-        icon: {
-          processing: false,
-          poster: '',
-          filename: '',
-          content_type: 'image/jpeg',
-          key: 'fd152d1004504c0ab68a99ce9e3fe5e7',
-          url: 'https://d2trgtv8344lrj.cloudfront.net/dist/1634904064/i/fd152d1004504c0ab68a99ce9e3fe5e7.jpeg?v=1704392507',
-        },
-        title: {
-          color: '#434761',
-          replacements: 'CleverTap Message Header',
-          text: 'CleverTap Message Header',
-        },
-        action: {
-          url: { ios: { replacements: '', text: '' }, android: { replacements: '', text: '' } },
-          links: [
-            {
-              kv: {},
-              url: {
-                ios: { replacements: 'https://valoraapp.com', text: 'https://valoraapp.com' },
-                android: { replacements: 'https://valoraapp.com', text: 'https://valoraapp.com' },
-              },
-              copyText: { replacements: 'https://valoraapp.com', text: 'https://valoraapp.com' },
-              text: 'CleverTap Message CTA',
-              bg: '#ffffff',
-              color: '#007bff',
-              type: 'url',
-            },
-          ],
-          hasLinks: true,
-          hasUrl: false,
-        },
-        message: {
-          color: '#434761',
-          replacements: 'CleverTap Message Body Text',
-          text: 'CleverTap Message Body Text',
-        },
-        media: {},
-        key: 99060129,
-      },
-    ],
-    enableTags: false,
-    custom_kv: [],
-    orientation: 'p',
-    bg: '#ffffff',
-  },
-  tags: [''],
-  isRead: true,
-}
