+++ src/home/selectors.test.ts
@@ -1,7 +1,7 @@
 import DeviceInfo from 'react-native-device-info'
-import { mockCleverTapInboxMessage } from 'src/home/cleverTapInbox'
 import { cleverTapInboxMessagesSelector, getExtraNotifications } from 'src/home/selectors'
 import { getMockStoreData } from 'test/utils'
+import { mockCleverTapInboxMessage } from 'test/values'
 
 describe(getExtraNotifications, () => {
   const mockedVersion = DeviceInfo.getVersion as jest.MockedFunction<typeof DeviceInfo.getVersion>
